Move proto files and update import paths
