+++ Makefile
@@ -4,5 +4,8 @@ buf-push:
 		echo "Error: VERSION is required. Usage: make buf-push VERSION=v0.7.2"; \
 		exit 1; \
 	fi
-	buf push --exclude-unnamed --label=main,$(VERSION)
-	cd testdata && buf dep update && buf generate && cd ..
+	cd protorelay
+	buf generate && buf push --exclude-unnamed --label=main,$(VERSION)
+	cd testdata
+	buf dep update && buf generate
+	cd ../..
