+++ protorelay/proto/relay/v1/relay.proto
@@ -2,7 +2,7 @@ syntax = "proto3";
 
 package relay.v1;
 
-option go_package = "github.com/theplant/relay/gen/relay/v1;relayv1";
+option go_package = "github.com/theplant/relay/protorelay/gen/relay/v1;relayv1";
 
 // OrderDirection specifies the direction of an order.
 enum OrderDirection {
